moved logic functions from main into logic file and imported functions, may need to clean up the main loop for cleaner input
